backend/requirements.txt
fastapi
uvicorn
sqlalchemy
psycopg2-binary
python-jose
passlib[bcrypt]
bcrypt
pydantic-settings
python-dotenv
pika
gunicorn
pydantic
email-validator
uvicorn
sqlalchemy-cockroachdb
redis